day4-7_fastapi/src/day5+6/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.0